refactor: organize poupeai app
