+++ apps/poupeai/models/__init__.py
@@ -1,18 +1,18 @@
 from .account import Account
 from .category import Category
-from .creditcard import CreditCard, Brand, Invoice
+from .credit_card import CreditCard, Brand, Invoice
 from .goal import Goal, GoalDeposit
 from .transaction import Transaction, AccountTransaction, CardTransaction
 
 __all__ = [
     'Account',
+    'Category',
+    'CreditCard',
+    'Brand',
+    'Invoice',
     'Goal',
     'GoalDeposit',
-    'Category',
     'Transaction',
-    'CardTransaction',
     'AccountTransaction',
-    'Brand',
-    'CreditCard',
-    'Invoice',
+    'CardTransaction',
 ]
