+++ apps/poupeai/models/transaction.py
@@ -2,7 +2,7 @@ from django.db import models
 from apps.authentication.models import CustomUser
 from .account import Account
 from .category import Category
-from .creditcard import CreditCard, Invoice
+from .credit_card import CreditCard, Invoice
 
 TRANSACTION_TYPES = (
     ('income', 'Receita'),
