+++ apps/poupeai/templates/poupeai/account_deletion.html
@@ -1,4 +1,4 @@
-{% extends "users_base.html" %}
+{% extends 'poupeai/poupeai_base.html' %}
 
 {% load static %}
