+++ apps/poupeai/templates/poupeai/poupeai_base.html
@@ -72,14 +72,14 @@
                   <p>Contas</p>
                 </a>
               </li>
-              <li class="nav-item {% if name == 'credit-card' %}active{% endif %}" class="nav-link">
-                <a href="{% url 'credit-card' %}">
+              <li class="nav-item {% if name == 'credit-cards' %}active{% endif %}" class="nav-link">
+                <a href="{% url 'credit-cards' %}">
                   <i class="fas fa-credit-card"></i>
                   <p>Cartões de Crédito</p>
                 </a>
               </li>
-              <li class="nav-item {% if name == 'my-goals' %}active{% endif %}" class="nav-link">
-                <a href="{% url 'my-goals' %}">
+              <li class="nav-item {% if name == 'goals' %}active{% endif %}" class="nav-link">
+                <a href="{% url 'goals' %}">
                   <i class="fas fa-bullseye"></i>
                   <p>Minhas Metas</p>
                 </a>
