+++ apps/poupeai/urls.py
@@ -7,11 +7,11 @@ urlpatterns = [
     path('categories/', CategoriesView.as_view(), name='categories'),
     path('transactions/', transactions_view, name='transactions'),
     path('accounts/', accounts_view, name='accounts'),
-    path('credit-card/', creditcard_view, name='credit-card'),
-    path('my-goals/', mygoals_view, name='my-goals'),
+    path('credit-cards/', creditcard_view, name='credit-cards'),
+    path('goals/', mygoals_view, name='goals'),
     path('help/', help_view, name='help'),
     path('profile/', profile_view, name='profile'),
-    path('account/deletion/', account_deletion_view, name='account-deletion'),
+    path('profile/deletion/', profile_deletion_view, name='profile-deletion'),
     path('user/admin/', admin_view, name='admin-dashboard'),
     path('notifications/', notifications_view, name='notifications'),
 ]
