+++ apps/poupeai/views/__init__.py
@@ -1,15 +1,15 @@
-from .accounts import accounts_view
-from .categories import CategoriesView
-from .creditcard import creditcard_view
-from .dashboard import dashboard_view
-from .help import help_view
-from .mygoals import mygoals_view
-from .profile import profile_view
-from .transactions import transactions_view
-from .account_deletion import account_deletion_view
+from .finance.accounts import accounts_view
+from .finance.categories import CategoriesView
+from .finance.credit_cards import creditcard_view
+from .core.dashboard import dashboard_view
+from .core.help import help_view
+from .finance.goals import mygoals_view
+from .profile.profiles import profile_view
+from .finance.transactions import transactions_view
+from .profile.profiles import profile_deletion_view
 from .admin_dashboard import admin_view
-from .home import home_view
-from .notifications import *
+from .core.home import home_view
+from .core.notifications import notifications_view
 
 __all__ = [
     'home_view',
@@ -21,7 +21,7 @@ __all__ = [
     'mygoals_view',
     'profile_view',
     'transactions_view',
-    'account_deletion_view',
+    'profile_deletion_view',
     'admin_view',
     'notifications_view',
 ]
