+++ apps/poupeai/views/admin_dashboard.py
@@ -2,4 +2,4 @@ from django.shortcuts import render
 
 def admin_view(request):
     context = {"name" : "admin-dashboard"}
-    return render(request, 'admin-dashboard.html', context)
+    return render(request, 'admin_dashboard.html', context)
