+++ apps/poupeai/views/core/dashboard.py
@@ -2,4 +2,4 @@ from django.shortcuts import render
 
 def dashboard_view(request):
     context = {"name" : "dashboard"}
-    return render(request, 'dashboard.html', context)
+    return render(request, 'poupeai/dashboard.html', context)
