+++ apps/poupeai/views/core/notifications.py
@@ -1,4 +1,4 @@
 from django.shortcuts import render
 
 def notifications_view(request):
-    return render(request, 'notifications.html')
+    return render(request, 'poupeai/notifications_page.html')
